tests/requirements_tests.txt
# Wymagania dla testów E2E Scenarzysta
pytest>=7.0.0
playwright>=1.40.0
python-docx>=0.8.11
openpyxl>=3.0.0
requests>=2.28.0
